requirements.txt
polars>=1.0.0
duckdb>=0.9.2
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
boto3>=1.18.0
botocore>=1.31.0
<<<<<<< HEAD
=======
streamlit>=1.30.0
plotly>=5.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
pydantic
aiohttp
>>>>>>> development
